NEW: Added SDLCursor class to hold the active cursor position
In the Editor the view system know which view is active and each view holds their own cursor. The active view will during rendering set the cursor to the graphics subsystem.

SDL concept of a text cursor is not the same as we have so it is implemented outside the context SDL