NEW: Cursor handling imported from SDL impl.
The cursor is handled by the windows. If we try to draw it when we update the actual windows it won't work. NCurses actually moves the cursor around when doing position-drawing (mvaddstr/ch/etc..) - and since the order of window drawing is not guranteed we have no clue which window is drawn first/last. The cursor needs to be moved in to the correct place AFTER all drawing is done. But must be chached during drawing for the active view...