NEW: Function to check if the textbuffer is editable
Currently we only check if we are parsing. As we can't really edit the buffer while parsing (we can to some extent, but there are a lot of corner cases).

This only (in practice) affects very large files during initialization. Otherwise it should have no noticeable effect on the performance.