CHG: Action Modifiers are now explicitly part of the actions and not just the keypress mask value.
In the config file we can declare optional modifiers to a keystroke like this:
NavigateLineDown = KeyDown + @SelectionModifier

This will internally create two actions - one with only key-down and one where the SelectionModifier is set to true.

And we used the mask to figure out which modifier might have been pressed and added that to the action. Now instead we do this when we build up the keymapping and storing that action explicitly.

The previous way could lead to false positives - like Shift+Something hadn't been declared as a 'SelectionModifier' - this would still be marked as such.