CHG: NewLine handling in EditController
This is a temporary fix and only works with certain languages (.C/.CPP) - need to generalize this in order to support all languages (who wants it).
The idea is that when you enter a new block or region of text that should be indented plus one we should create a new line and position the cursor at the newly indented position. This depends on the language parser to detect what constitutes a block.