FIX: Moved cmdlet_prefix to main configuration
The cmdlet_prefix was declared in several places and for various purposes and they settings would interchange in the code-path. Checked in from one part of the config and parsed using another - ergo - not good.

The drawback is that the prefix is the same in all places - which is also a good thing.