CHG: Handling of new document creation and activation
When creating a new document the parent of the document should not be an already existing document. Instead we go back one step and take the parent folder (or root workspace).

Also on LoadData we check if the file exists before trying to load. This prohibits trying to load (and fail) new documents.