Problem with language tokenizer and whole-word tokens.
Currently we detected tokens like 'ifelsewhilefor' as separate keywords; [if, else, while, for] but this is clearly wrong. This commit contains the test case showing this problem.

I need to split token identification/classification in two stages - whole-tokens and attached-tokens. Currently all are treated as 'attached' i.e. they can be attached to something.

LangLineTokenizer has 'fixme' comments where we need to do stuff. But also need to add a flag to the identifier structure.