FIXES: was abusing iterators
Hunting for a memory overwrite bug (found by accident).
I was using the () operator of an iterator to gain access to the underlying data. This has been replaced to properly use the string and iterators for start/end.

Also replaced std::u32string for std::u32string_view when I actually just use it for checking - no need to create and copy a whole object.